# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  domain.LBStatus:
    enum:
    - CREATING
    - ACTIVE
    - DRAINING
    - DELETED
    type: string
    x-enum-varnames:
    - LBStatusCreating
    - LBStatusActive
    - LBStatusDraining
    - LBStatusDeleted
  domain.LBTarget:
    properties:
      health:
        description: '"healthy" | "unhealthy" | "unknown"'
        type: string
      id:
        type: string
      instance_id:
        type: string
      lb_id:
        type: string
      port:
        type: integer
      weight:
        type: integer
    type: object
  domain.LoadBalancer:
    properties:
      algorithm:
        description: '"round-robin" | "least-conn"'
        type: string
      created_at:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
      name:
        type: string
      port:
        type: integer
      status:
        $ref: '#/definitions/domain.LBStatus'
      version:
        type: integer
      vpc_id:
        type: string
    type: object
...
  domain.Object:
    properties:
      arn:
        type: string
      bucket:
        type: string
      content_type:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      key:
        type: string
      size_bytes:
        type: integer
    type: object
...
  domain.Volume:
    properties:
      created_at:
        type: string
      id:
        type: string
      instance_id:
        type: string
      mount_path:
        type: string
      name:
        type: string
      size_gb:
        type: integer
      status:
        $ref: '#/definitions/domain.VolumeStatus'
      updated_at:
        type: string
    type: object
  domain.VolumeStatus:
    enum:
    - AVAILABLE
    - IN-USE
    - DELETING
    type: string
    x-enum-varnames:
    - VolumeStatusAvailable
    - VolumeStatusInUse
    - VolumeStatusDeleting
  httphandlers.AddTargetRequest:
    properties:
      instance_id:
        type: string
      port:
        type: integer
      weight:
        type: integer
    required:
    - instance_id
    - port
    type: object
...
  httphandlers.CreateLBRequest:
    properties:
      algorithm:
        type: string
      name:
        type: string
      port:
        type: integer
      vpc_id:
        type: string
    required:
    - name
    - port
    - vpc_id
    type: object
  httphandlers.CreateVolumeRequest:
    properties:
      name:
        type: string
      size_gb:
        type: integer
    required:
    - name
    type: object
...
  httputil.Meta:
    properties:
      request_id:
        type: string
      timestamp:
        type: string
    type: object
  httputil.Response:
    properties:
      data: {}
      error: {}
      meta:
        $ref: '#/definitions/httputil.Meta'
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Mini AWS Compute API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Mini AWS API
  version: "1.0"
paths:
...
  /lb:
    get:
      description: Gets a list of all load balancers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.LoadBalancer'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List all load balancers
      tags:
      - loadbalancers
    post:
      consumes:
      - application/json
      description: Creates a new load balancer in a VPC
      parameters:
      - description: LB creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httphandlers.CreateLBRequest'
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/domain.LoadBalancer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new load balancer
      tags:
      - loadbalancers
  /lb/{id}:
    delete:
      description: Removes a load balancer and stops associated proxy
      parameters:
      - description: LB ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete a load balancer
      tags:
      - loadbalancers
    get:
      description: Gets detailed information about a specific load balancer
      parameters:
      - description: LB ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.LoadBalancer'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Get load balancer details
      tags:
      - loadbalancers
  /lb/{id}/targets:
    get:
      description: Gets a list of targets for a load balancer
      parameters:
      - description: LB ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.LBTarget'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List all targets for a load balancer
      tags:
      - loadbalancers
    post:
      consumes:
      - application/json
      description: Registers a compute instance to receive traffic from the load balancer
      parameters:
      - description: LB ID
        in: path
        name: id
        required: true
        type: string
      - description: Target details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httphandlers.AddTargetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httputil.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Add a target to a load balancer
      tags:
      - loadbalancers
  /lb/{id}/targets/{instanceId}:
    delete:
      description: Deregisters a compute instance from the load balancer
      parameters:
      - description: LB ID
        in: path
        name: id
        required: true
        type: string
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Remove a target from a load balancer
      tags:
      - loadbalancers
  /storage/{bucket}:
    get:
      description: Gets a list of all objects within a specific bucket
      parameters:
      - description: Bucket name
        in: path
        name: bucket
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Object'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: List objects in a bucket
      tags:
      - storage
  /storage/{bucket}/{key}:
    delete:
      description: Removes an object from the specified bucket
      parameters:
      - description: Bucket name
        in: path
        name: bucket
        required: true
        type: string
      - description: Object key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete an object
      tags:
      - storage
    get:
      description: Streams the specified object as an attachment
      parameters:
      - description: Bucket name
        in: path
        name: bucket
        required: true
        type: string
      - description: Object key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Object content
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Download an object
      tags:
      - storage
    put:
      consumes:
      - application/octet-stream
      description: Uploads a file/object to the specified bucket and key
      parameters:
      - description: Bucket name
        in: path
        name: bucket
        required: true
        type: string
      - description: Object key
        in: path
        name: key
        required: true
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Object'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Upload an object
      tags:
      - storage
  /volumes:
    get:
      description: Gets a list of all existing block storage volumes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Volume'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: List all volumes
      tags:
      - volumes
    post:
      consumes:
      - application/json
      description: Creates a new block storage volume
      parameters:
      - description: Volume creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httphandlers.CreateVolumeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Volume'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new volume
      tags:
      - volumes
  /volumes/{id}:
    delete:
      description: Removes a block storage volume
      parameters:
      - description: Volume ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete a volume
      tags:
      - volumes
    get:
      description: Gets detailed information about a specific block storage volume
      parameters:
      - description: Volume ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Volume'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Get volume details
      tags:
      - volumes
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"